socialLogin components create and inside google facebook and github login method create that use signIn and signUp Form inside
